New Feature: Se creo la opción de cerrar sesión, además se genera el rol usuario y administrador independientemente, y se asigna restricciones a usuario: Estudiante

+++ src/app/pages/user/principalComponent/user-component/user-component.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Store } from '@ngrx/store';
-import { Observable, Subscriber, Subscription, from } from 'rxjs';
+import { Observable, Subscriber, Subscription, from, of } from 'rxjs';
 import { SemestreDataService } from 'src/app/data/material/material-data/semestre-data.service';
 import { StudentDataService } from 'src/app/data/user/student-data/student-data.service';
 import { StudentJsonDataService } from 'src/app/data/user/student-data/student-json-data.service';
@@ -12,6 +12,8 @@ import { ObjDataStudent } from 'src/app/utilities/interfases/obj-dataStudent';
 import { ObjSemestre } from 'src/app/utilities/interfases/obj-semestre';
 import { ObjStudent } from 'src/app/utilities/interfases/obj-student';
 import { ObjUsers } from 'src/app/utilities/interfases/obj-users';
+import { AuthService } from 'src/app/utilities/services/auth/auth.service';
+import { LogueoService } from 'src/app/utilities/services/logueo/logueo.service';
 import { AppState } from 'src/app/utilities/store/app.reducer';
 import { loadIdUser } from 'src/app/utilities/store/user/user.action';
 
@@ -23,6 +25,8 @@ import { loadIdUser } from 'src/app/utilities/store/user/user.action';
 export class UserComponentComponent implements OnInit {
   // private datastudents = new UserDataModule();
   public student!: ObjUsers[];
+  private user!: ObjUsers;
+  private userId: string = '';
   public semestre!: ObjSemestre;
   private dataJSON!: ObjDataStudent[];
   private suscription!: Subscription;
@@ -36,6 +40,8 @@ export class UserComponentComponent implements OnInit {
   constructor(private serviceDataStudent: StudentDataService,
     private dataJSONStudent: StudentJsonDataService,
     private userDataService: UserJsonDataService,
+    private authUserService: AuthService,
+    private logSessionServices: LogueoService,
     private router: Router,
     private store: Store<AppState>) { }
 
@@ -49,30 +55,55 @@ export class UserComponentComponent implements OnInit {
     //     console.log(this.student)
     // });
     this.serviceDataStudent.resetDataStudent([]);
-    this.suscription = this.userDataService.getDataUsers().subscribe(
-      (r: ObjUsers[]) => {
-        from(r).subscribe(     //Se recorre el objeto traido del API
-          {
-            next: value => {
-              if (value.rol === this.rol) {
-                this.serviceDataStudent.setDataStudent.subscribe(
-                (obj: ObjUsers[]) => {
-                  console.log('ArrayStudent', this.serviceDataStudent.getStudent()),
-                  console.log('USER COMPONENT value'),
-                  console.log(value),
-                  console.log('USER COMPONENT obj'),
-                  console.log(obj),
+    if (this.authUserService.getSession()) {
+      this.suscription = this.userDataService.getDataUsers().subscribe(
+        (r: ObjUsers[]) => {
+          from(r).subscribe(     //Se recorre el objeto traido del API
+            {
+              next: value => {
+                if (value.rol === this.rol) {
+                  this.serviceDataStudent.setDataStudent.subscribe(
+                    (obj: ObjUsers[]) => {
+                      console.log('ArrayStudent', this.serviceDataStudent.getStudent()),
+                        console.log('USER COMPONENT value'),
+                        console.log(value),
+                        console.log('USER COMPONENT obj'),
+                        console.log(obj),
+                        obj.push(value),
+                        console.log(obj),
+                        console.log('ArrayStudent', this.serviceDataStudent.getStudent()),
+                        this.student = obj
+                    });
+                }
+              },
+              error: err => console.error('ERROR: ', err),
+              complete: () => console.info('ok')
+            });
+        });
+    }
+    this.userId = this.logSessionServices.getInfoSession().get('id')?.value;
+    this.suscription = this.userDataService.getDataUserID(this.userId).subscribe( 
+      (r: ObjUsers) => 
+      {of(r).subscribe(
+        {
+          next: value => {
+            this.serviceDataStudent.setDataStudent.subscribe(
+              (obj: ObjUsers[]) => {
                   obj.push(value),
-                  console.log(obj),
-                  console.log('ArrayStudent', this.serviceDataStudent.getStudent()),
                   this.student = obj
-                });
-              }
-            },
-            error: err => console.error('ERROR: ', err),
-            complete: () => console.info('ok')
-          });
+              });
+          },
+          error: err => console.log('err: ', err),
+          complete: () => console.log('ok'),
+        }
+      );
+        // this.user = r,
+        // console.log('Student'),
+        // console.log(this.user),
+        // this.student.push(this.user)
       });
+     
+ 
 
     //  this.serviceDataStudent.getStudent.subscribe(
     //   (r) => this.student = r
@@ -97,8 +128,13 @@ export class UserComponentComponent implements OnInit {
   ];
 
   delete(id: string) {
-    this.userDataService.deleteDataUser(id).subscribe();
-    this.router.navigate(['inicio']);
+    if (this.authUserService.getSession()) {
+      this.userDataService.deleteDataUser(id).subscribe();
+      this.router.navigate(['inicio']);
+    }else{
+      console.log('BORRAR COMPONENT');
+      this.router.navigate(['usuarios/borrar']);
+    }
   }
 
   modify(id: string) {
@@ -120,7 +156,7 @@ export class UserComponentComponent implements OnInit {
     this.store.dispatch(loadIdUser({ data: this.dataJSONId }));
   }
 
-  
+
 }
